[WIP] introduce a container launcher
This launcher covers everything that the k8s and docker launchers do,
without some of their bugs.  In particular, it

 - handles file configuration for containers under kubernetes
 - handles changes to pod annotations
 - correctly logs from file or socket based on
   `logs_config.{k8s,docker}_container_use_file`
 - captures logs under containerd when `k8s_container_use_file` is false

Outside of the situations where they have obvious bugs, this launcher
attempts to preserve behavior of the kubernetes and docker launchers in
the details, such as the "Source" and "Service" properties.  It does not
preserve those launchers' behavior with respect to `agent status`
output, as that behavior was not very intuitive.

In this commit, it is not connected to anything; that will come later.

NOTES:

The following logic matches existing non-buggy behavior:

if running in k8s:
  if kcuf: log from files in /var/log/pods
  else: log from socket (maybe: verify socket works on containerd?)
else:
  if dcuf: log from files in /var/lib/containers
  else: log from socket

WIP:
 - need to sort out build flags -- should this whole thing be +docker?
 - need to wrap external tailers in an internal containerTailer interface?
 - probably move tailerfactory to /internal/